requirements.txt
# Find new packages at https://pypi.org/search/
Django==4.0.3
MarkupSafe==1.1.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
selenium==4.1.3
djangorestframework==3.13.1
Markdown==3.3.6
django-filter==21.1
django-oauth-toolkit==1.7.1

# Potentially not needed
sqlparse==0.3.0
whitenoise==6.0.0
Pillow==9.1.0
